+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+
+
+## v0.5.7
 ### Added
 - make sort direction of event overview persistent.
 - member statistic for attendance poll statistics.
+- added event start/end date update function from edit to create form too.
 
 ## v0.5.6
 ### Added